backend/api-service/requirements.txt
# User Management API Requirements
# Core Web Framework
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
python-multipart==0.0.20
email-validator==2.3.0

# Firebase
firebase-admin==6.1.0
google-cloud-firestore==2.13.1 
